implement summary API
