+++ summary_be/app.py
@@ -3,6 +3,8 @@ from flask import Flask
 from flask_cors import CORS
 import dotenv
 
+from routes.summarization import summarization_route
+
 basedir = os.path.abspath(os.path.dirname(__file__))
 secret_file = os.path.join(basedir, ".env")
 if os.path.isfile(secret_file):
@@ -13,6 +15,8 @@ CORS(app)
 PORT = "8000"
 if 'PORT' in os.environ:
     PORT = os.environ['PORT']
+
+app.register_blueprint(summarization_route)
 
 if __name__ == '__main__':
     app.run(port=PORT, debug=True)
