--- summary_be/utils/utils.py
@@ -1,1 +1,0 @@
-# placeholder
